+++ src/Filter/ClassFilter.php
@@ -49,7 +49,10 @@ class ClassFilter extends Filter
 
     public function getDefaultOptions()
     {
-        return [];
+        return [
+            'operator_type' => EqualType::class,
+            'operator_options' => [],
+        ];
     }
 
     public function getFieldType()
@@ -74,7 +77,8 @@ class ClassFilter extends Filter
     public function getRenderSettings()
     {
         return [DefaultType::class, [
-            'operator_type' => EqualType::class,
+            'operator_type' => $this->getOption('operator_type'),
+            'operator_options' => $this->getOption('operator_options'),
             'field_type' => $this->getFieldType(),
             'field_options' => $this->getFieldOptions(),
             'label' => $this->getLabel(),