+++ tests/Model/ModelManagerTest.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Sonata\DoctrineORMAdminBundle\Tests\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Persistence\ManagerRegistry;
 use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
@@ -21,6 +22,7 @@ use Doctrine\DBAL\Platforms\PostgreSqlPlatform;
 use Doctrine\DBAL\Types\Type;
 use Doctrine\ORM\Configuration;
 use Doctrine\ORM\EntityManager;
+use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\Mapping\ClassMetadataFactory;
 use Doctrine\ORM\OptimisticLockException;
@@ -69,9 +71,23 @@ class ModelManagerTest extends TestCase
         }
     }
 
-    public function testSortParameters(): void
+    public function testInstantiateWithDeprecatedRegistryInterface(): void
     {
         $registry = $this->createMock(RegistryInterface::class);
+        $manager = new ModelManager($registry);
+        $em = $this->createMock(EntityManagerInterface::class);
+
+        $registry->expects($this->once())
+            ->method('getManagerForClass')
+            ->with('x')
+            ->willReturn($em)
+        ;
+        $this->assertSame($em, $manager->getEntityManager('x'));
+    }
+
+    public function testSortParameters(): void
+    {
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $manager = new ModelManager($registry);
 
@@ -416,7 +432,7 @@ class ModelManagerTest extends TestCase
             ->method('getConnection')
             ->willReturn($conn);
 
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
         $registry->expects($this->any())
             ->method('getManagerForClass')
             ->willReturn($em);
@@ -466,7 +482,7 @@ class ModelManagerTest extends TestCase
             ->method('getConnection')
             ->willReturn($conn);
 
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
         $registry->expects($this->any())
             ->method('getManagerForClass')
             ->willReturn($em);
@@ -511,7 +527,7 @@ class ModelManagerTest extends TestCase
             ->method('getConnection')
             ->willReturn($conn);
 
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
         $registry->expects($this->any())
             ->method('getManagerForClass')
             ->willReturn($em);
@@ -607,7 +623,7 @@ class ModelManagerTest extends TestCase
 
         $metadataFactory = $this->createMock(ClassMetadataFactory::class);
         $modelManager = $this->createMock(ObjectManager::class);
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $classMetadata = new ClassMetadata($class);
         $classMetadata->reflClass = new \ReflectionClass($class);
@@ -638,7 +654,7 @@ class ModelManagerTest extends TestCase
 
     public function testCollections(): void
     {
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
         $model = new ModelManager($registry);
 
         $collection = $model->getModelCollectionInstance('whyDoWeEvenHaveThisParameter');
@@ -662,7 +678,7 @@ class ModelManagerTest extends TestCase
 
     public function testModelTransform(): void
     {
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
         $model = new ModelManager($registry);
 
         $result = $model->modelTransform('thisIsNotUsed', 'doWeNeedThisMethod');
@@ -674,7 +690,7 @@ class ModelManagerTest extends TestCase
     {
         $datagrid = $this->createMock(DatagridInterface::class);
         $filter = $this->createMock(FilterInterface::class);
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $datagrid->expects($this->once())
             ->method('getValues')
@@ -694,7 +710,7 @@ class ModelManagerTest extends TestCase
 
     public function testGetModelInstanceException(): void
     {
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $model = new ModelManager($registry);
 
@@ -705,7 +721,7 @@ class ModelManagerTest extends TestCase
 
     public function testGetModelInstanceForProtectedEntity(): void
     {
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $model = new ModelManager($registry);
 
@@ -714,7 +730,7 @@ class ModelManagerTest extends TestCase
 
     public function testGetEntityManagerException(): void
     {
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $model = new ModelManager($registry);
 
@@ -725,7 +741,7 @@ class ModelManagerTest extends TestCase
 
     public function testGetNewFieldDescriptionInstanceException(): void
     {
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $model = new ModelManager($registry);
 
@@ -739,7 +755,7 @@ class ModelManagerTest extends TestCase
      */
     public function testCreate($exception): void
     {
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $entityManger = $this->createMock(EntityManager::class);
 
@@ -778,7 +794,7 @@ class ModelManagerTest extends TestCase
      */
     public function testUpdate($exception): void
     {
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $entityManger = $this->createMock(EntityManager::class);
 
@@ -805,7 +821,7 @@ class ModelManagerTest extends TestCase
      */
     public function testRemove($exception): void
     {
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $entityManger = $this->createMock(EntityManager::class);
 
@@ -829,7 +845,7 @@ class ModelManagerTest extends TestCase
 
     public function testFindBadId(): void
     {
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $model = new ModelManager($registry);
 
@@ -838,7 +854,7 @@ class ModelManagerTest extends TestCase
 
     public function testGetUrlsafeIdentifierException(): void
     {
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $model = new ModelManager($registry);
 
@@ -849,7 +865,7 @@ class ModelManagerTest extends TestCase
 
     public function testGetUrlsafeIdentifierNull(): void
     {
-        $registry = $this->createMock(RegistryInterface::class);
+        $registry = $this->createMock(ManagerRegistry::class);
 
         $model = new ModelManager($registry);